requirements.txt
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
streamlit==1.29.0
openai==0.28
pillow